requirements.txt
anthropic>=0.94.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
openrgb-python>=0.3.6
